refactor(mvc): legacy mvc 제거 및 모든 컨트롤러를 어노테이션 기반으로 변경
