+++ README.md
@@ -39,8 +39,8 @@
   - [ ] model에 데이터가 1개면 값을 그대로 반환한다.
   - [ ] model에 데이터가 2개 이상이면 Map형태 그대로 JSON으로 변환해서 반환한다. 
   - [ ] UserController가 JSON 형태로 응답을 반환한다.
-- [ ] Legacy MVC를 제거한다.
-  - [ ] 모든 컨트롤러를 어노테이션 기반 MVC로 변경한다.
-  - [ ] asis 패키지의 모든 클래스를 삭제해도 프로그램이 정상적으로 동작해야한다.
+- [X] Legacy MVC를 제거한다.
+  - [X] 모든 컨트롤러를 어노테이션 기반 MVC로 변경한다.
+  - [X] asis 패키지의 모든 클래스를 삭제해도 프로그램이 정상적으로 동작해야한다.
 - [ ] HTTP Request Body로 JSON 타입의 데이터를 받았을 때, 이를 자바 객체로 변환한다.
   - [ ] JSON을 자바 객체로 변환할 때 Jackson 라이브러리를 이용한다.
