+++ mvc/src/main/java/nextstep/mvc/DispatcherServlet.java
@@ -6,8 +6,8 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.util.Map;
 import java.util.Optional;
-import nextstep.mvc.controller.tobe.HandlerAdapterRegistry;
-import nextstep.mvc.controller.tobe.HandlerMappingRegistry;
+import nextstep.mvc.controller.HandlerAdapterRegistry;
+import nextstep.mvc.controller.HandlerMappingRegistry;
 import nextstep.mvc.exception.HandlerAdapterException;
 import nextstep.mvc.exception.HandlerMappingException;
 import nextstep.mvc.exception.ViewException;
