+++ mvc/src/main/java/nextstep/mvc/controller/HandlerAdapterRegistry.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.controller.tobe;
+package nextstep.mvc.controller;
 
 import java.util.ArrayList;
 import java.util.List;
